ai resume screening agent/AI-Resume-Agent/requirements.txt
streamlit>=1.20.0
sentence-transformers>=2.2.2
spacy>=3.5.0
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.23.0
PyPDF2>=3.0.0
python-docx>=0.8.11
regex>=2022.10.31
joblib>=1.2.0
